+++ README.md
@@ -14,7 +14,7 @@ Comprehensive solenoid lens design tool.
  - ASTRA interface
 
 **Dependencies**:
- - numpy, pandas, scipy
+ - numpy, pandas, scipy, f90nml
  - ASTRA
  - Has not been tested/configured to work with Windows, i.e. linux only
